build: update vs sln
